move tags from ctx to tag.ex for #68
