+++ lib/app/item.ex
@@ -10,7 +10,7 @@ defmodule App.Item do
     # an item always belongs to a person
     field :person_id, :id
     field :status, :id
-    many_to_many :tags, App.Ctx.Tag, join_through: "item_tags"
+    many_to_many :tags, App.Tag, join_through: "item_tags"
     timestamps()
   end
