+++ lib/app_web/controllers/tag_controller.ex
@@ -1,21 +1,19 @@
 defmodule AppWeb.TagController do
   use AppWeb, :controller
-
-  alias App.Ctx
-  alias App.Ctx.Tag
+  alias App.Tag
 
   def index(conn, _params) do
-    tags = Ctx.list_tags()
+    tags = Tag.list_tags()
     render(conn, "index.html", tags: tags)
   end
 
   def new(conn, _params) do
-    changeset = Ctx.change_tag(%Tag{})
+    changeset = Tag.change_tag(%Tag{})
     render(conn, "new.html", changeset: changeset)
   end
 
   def create(conn, %{"tag" => tag_params}) do
-    case Ctx.create_tag(tag_params) do
+    case Tag.create_tag(tag_params) do
       {:ok, tag} ->
         conn
         |> put_flash(:info, "Tag created successfully.")
@@ -27,20 +25,20 @@ defmodule AppWeb.TagController do
   end
 
   def show(conn, %{"id" => id}) do
-    tag = Ctx.get_tag!(id)
+    tag = Tag.get_tag!(id)
     render(conn, "show.html", tag: tag)
   end
 
   def edit(conn, %{"id" => id}) do
-    tag = Ctx.get_tag!(id)
-    changeset = Ctx.change_tag(tag)
+    tag = Tag.get_tag!(id)
+    changeset = Tag.change_tag(tag)
     render(conn, "edit.html", tag: tag, changeset: changeset)
   end
 
   def update(conn, %{"id" => id, "tag" => tag_params}) do
-    tag = Ctx.get_tag!(id)
+    tag = Tag.get_tag!(id)
 
-    case Ctx.update_tag(tag, tag_params) do
+    case Tag.update_tag(tag, tag_params) do
       {:ok, tag} ->
         conn
         |> put_flash(:info, "Tag updated successfully.")
@@ -52,8 +50,8 @@ defmodule AppWeb.TagController do
   end
 
   def delete(conn, %{"id" => id}) do
-    tag = Ctx.get_tag!(id)
-    {:ok, _tag} = Ctx.delete_tag(tag)
+    tag = Tag.get_tag!(id)
+    {:ok, _tag} = Tag.delete_tag(tag)
 
     conn
     |> put_flash(:info, "Tag deleted successfully.")
