+++ test/app/ctx_test.exs
@@ -3,65 +3,6 @@ defmodule App.CtxTest do
 
   alias App.Ctx
 
-  describe "tags" do
-    alias App.Ctx.Tag
-
-    @valid_attrs %{text: "some text"}
-    @update_attrs %{text: "some updated text"}
-    @invalid_attrs %{text: nil}
-
-    def tag_fixture(attrs \\ %{}) do
-      {:ok, tag} =
-        attrs
-        |> Enum.into(@valid_attrs)
-        |> Ctx.create_tag()
-
-      tag
-    end
-
-    test "list_tags/0 returns all tags" do
-      tag = tag_fixture()
-      assert Ctx.list_tags() == [tag]
-    end
-
-    test "get_tag!/1 returns the tag with given id" do
-      tag = tag_fixture()
-      assert Ctx.get_tag!(tag.id) == tag
-    end
-
-    test "create_tag/1 with valid data creates a tag" do
-      assert {:ok, %Tag{} = tag} = Ctx.create_tag(@valid_attrs)
-      assert tag.text == "some text"
-    end
-
-    test "create_tag/1 with invalid data returns error changeset" do
-      assert {:error, %Ecto.Changeset{}} = Ctx.create_tag(@invalid_attrs)
-    end
-
-    test "update_tag/2 with valid data updates the tag" do
-      tag = tag_fixture()
-      assert {:ok, %Tag{} = tag} = Ctx.update_tag(tag, @update_attrs)
-      assert tag.text == "some updated text"
-    end
-
-    test "update_tag/2 with invalid data returns error changeset" do
-      tag = tag_fixture()
-      assert {:error, %Ecto.Changeset{}} = Ctx.update_tag(tag, @invalid_attrs)
-      assert tag == Ctx.get_tag!(tag.id)
-    end
-
-    test "delete_tag/1 deletes the tag" do
-      tag = tag_fixture()
-      assert {:ok, %Tag{}} = Ctx.delete_tag(tag)
-      assert_raise Ecto.NoResultsError, fn -> Ctx.get_tag!(tag.id) end
-    end
-
-    test "change_tag/1 returns a tag changeset" do
-      tag = tag_fixture()
-      assert %Ecto.Changeset{} = Ctx.change_tag(tag)
-    end
-  end
-
   describe "status" do
     alias App.Ctx.Status
